# Type system performance improvements