On a sample dataset of ~500k rows, achieves a > 90% reduction in running time of the type inference logic in `import_rows()` on an M3 Max CPU, from 32.1 seconds to 2.56 seconds.

Most of the reduction (from 32 -> ~6 seconds) is achieved by eliminating the use of `deepcopy()` in `ColumnType.copy()`. The remainder is achieved by reducing object creation and dynamic evaluation in `ColumnType` and subclasses.

It is assumed that `ColumnType` instances are immutable. I believe that further optimizations are possible by making `ColumnType` instances formally immutable (as frozen dataclasses, say) and minimizing object creation. But those changes would be more invasive, and I'm pretty sure this is enough to get type inference off the critical performance path.

UPDATED: Optimized `JsonType._validate_literal`, reducing the time to 2.3 seconds (93% reduction vs. baseline)