@aaron-siegel 2024-08-26
<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/pixeltable/pixeltable/264)
<!-- Reviewable:end -->